avoid using deprecated  and
